third_party/blink/web_tests/http/tests/devtools/throttling/mobile-throttling-expected.txt

Tests that mobile, network, and CPU throttling interact with each other logically.

Initial throttling state
=== THROTTLING STATE ===
Network throttling - download: -1 upload: -1 latency: 0
CPU throttling rate: 1
Device mode throttling: No throttling
Network panel throttling: No throttling
Network conditions drawer throttling: No throttling
Performance panel network throttling: No throttling
Performance panel CPU throttling: No throttling
========================

Change to offline in device mode
=== THROTTLING STATE ===
Network throttling - download: 0 upload: 0 latency: 0
CPU throttling rate: 1
Device mode throttling: Offline
Network panel throttling: Offline
Network conditions drawer throttling: Offline
Performance panel network throttling: Offline
Performance panel CPU throttling: No throttling
========================

Change to low-end mobile in device mode
=== THROTTLING STATE ===
Network throttling - download: 50000 upload: 50000 latency: 2000
CPU throttling rate: 6
Device mode throttling: Low-end mobile
Network panel throttling: 3G
Network conditions drawer throttling: 3G
Performance panel network throttling: 3G
Performance panel CPU throttling: 6× slowdown
========================

Change network to 3G
=== THROTTLING STATE ===
Network throttling - download: 50000 upload: 50000 latency: 2000
CPU throttling rate: 6
Device mode throttling: Low-end mobile
Network panel throttling: 3G
Network conditions drawer throttling: 3G
Performance panel network throttling: 3G
Performance panel CPU throttling: 6× slowdown
========================

Change network to slow 4G
=== THROTTLING STATE ===
Network throttling - download: 180000 upload: 84375 latency: 562.5
CPU throttling rate: 6
Device mode throttling: Custom
Network panel throttling: Slow 4G
Network conditions drawer throttling: Slow 4G
Performance panel network throttling: Slow 4G
Performance panel CPU throttling: 6× slowdown
========================

Change network to fast 4G
=== THROTTLING STATE ===
Network throttling - download: 1012500 upload: 168750 latency: 165
CPU throttling rate: 6
Device mode throttling: Custom
Network panel throttling: Fast 4G
Network conditions drawer throttling: Fast 4G
Performance panel network throttling: Fast 4G
Performance panel CPU throttling: 6× slowdown
========================

Change to mid-tier mobile in device mode
=== THROTTLING STATE ===
Network throttling - download: 180000 upload: 84375 latency: 562.5
CPU throttling rate: 4
Device mode throttling: Mid-tier mobile
Network panel throttling: Slow 4G
Network conditions drawer throttling: Slow 4G
Performance panel network throttling: Slow 4G
Performance panel CPU throttling: 4× slowdown
========================

Change CPU throttling to low-end mobile
=== THROTTLING STATE ===
Network throttling - download: 180000 upload: 84375 latency: 562.5
CPU throttling rate: 6
Device mode throttling: Custom
Network panel throttling: Slow 4G
Network conditions drawer throttling: Slow 4G
Performance panel network throttling: Slow 4G
Performance panel CPU throttling: 6× slowdown
========================

Change CPU throttling to mid-tier mobile
=== THROTTLING STATE ===
Network throttling - download: 180000 upload: 84375 latency: 562.5
CPU throttling rate: 4
Device mode throttling: Mid-tier mobile
Network panel throttling: Slow 4G
Network conditions drawer throttling: Slow 4G
Performance panel network throttling: Slow 4G
Performance panel CPU throttling: 4× slowdown
========================

Change to no throttling in device mode
=== THROTTLING STATE ===
Network throttling - download: -1 upload: -1 latency: 0
CPU throttling rate: 1
Device mode throttling: No throttling
Network panel throttling: No throttling
Network conditions drawer throttling: No throttling
Performance panel network throttling: No throttling
Performance panel CPU throttling: No throttling
========================
